Added checks when deleting/creating types

+++ Web Application/src/main/java/com/example/demo/controllers/TypeController.java
@@ -8,6 +8,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import java.util.List;
 
@@ -29,9 +30,16 @@ public class TypeController {
         return "home/types/create";
     }
     @PostMapping("/create")
-    public String addType(@ModelAttribute Type type){
-        typeService.add(type);
-        return "redirect:/types/list";
+    public String addType(@ModelAttribute Type type, RedirectAttributes redirectAttributes){
+        if(typeService.findById(type.getType()) == null){
+            typeService.add(type);
+            return "redirect:/types/list";
+        }
+        String failedMessage = String.format("Unable to add type - type with name '%s' already exists", type.getType());
+        redirectAttributes.addFlashAttribute("message", failedMessage);
+        redirectAttributes.addFlashAttribute("alertClass", "alert-danger");
+
+        return "redirect:/types/create";
     }
 
     @GetMapping("/view-one/{id}")
@@ -42,18 +50,22 @@ public class TypeController {
 
     @GetMapping("/edit/{id}")
     public String editType(@PathVariable("id") String type, Model model) {
-        System.out.println("Get mapping edit type: " + type);
         model.addAttribute("type", typeService.findById(type));
         return "home/types/edit";
     }
     @PostMapping("/update")
     public String updateType(@ModelAttribute Type type) {
-        System.out.println("Post mapping edit type: " + type.getType());
         typeService.update(type.getType(), type);
         return "redirect:/types/list";
     }
     @GetMapping("/delete/{id}")
-    public String deleteType(@PathVariable("id") String type) {
+    public String deleteType(@PathVariable("id") String type, RedirectAttributes redirectAttributes) {
+        if (typeService.usedInMotorHome(type)) {
+            String failedMessage = "Unable to delete - type is already used in one or more motorhomes";
+            redirectAttributes.addFlashAttribute("message", failedMessage);
+            redirectAttributes.addFlashAttribute("alertClass", "alert-danger");
+            return "redirect:/types/list";
+        }
         typeService.delete(type);
         return "redirect:/types/list";
     }
